Decouple box.js and fix imports
