+++ js/viz/objects/ReconstructedParticle.js
@@ -1,6 +1,6 @@
 import { EDMObject } from "./EDMObject.js";
 import { parseCharge } from "../lib/parseCharge.js";
-import { addLinesToBox } from "../draw/box.js";
+import { addLinesToBox } from "../draw/font.js";
 import { objectColor } from "../constants/vizStyles.js";
 
 class ReconstructedParticle extends EDMObject {
